+++ shuup/admin/modules/manufacturers/views/edit.py
@@ -80,7 +80,7 @@ class ManufacturerEditView(CreateOrUpdateView):
     def get_toolbar(self):
         object = self.get_object()
         delete_url = (
-            reverse_lazy("shuup_admin:manufacture.delete", kwargs={"pk": object.pk})
+            reverse_lazy("shuup_admin:manufacturer.delete", kwargs={"pk": object.pk})
             if object.pk else None)
         return get_default_edit_toolbar(self, self.get_save_form_id(), delete_url=delete_url)