+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 List all changes after the last release here (newer on top). Each change on a separate bullet point line.
 ### Added
 
-- Admin: Added option to delete manufacture
+- Admin: Added option to delete manufacturer
 
 ## [1.10.10] - 2020-03-24